requirements.txt
cdsapi==0.7.5
cfgrib==0.9.15.0
pandas==2.3.1
Pillow==11.3.0
pytesseract==0.3.13
Requests==2.32.4
tqdm==4.66.5
xarray==2025.3.0
